Fix issues (move from BigDecimal to Double)

+++ src/main/java/fr/polytech/cloud/entities/PositionMongoDBEntity.java
@@ -2,12 +2,11 @@ package fr.polytech.cloud.entities;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public @Data class PositionMongoDBEntity {
 
     private String type;
 
-    private List<BigDecimal> coordinates;
+    private List<Double> coordinates;
 }
